rename existing db to db_inmemory
