+++ tests/test_route_users.py
@@ -40,38 +40,21 @@ def test_create_new_user_missing_password(client: Client):
 
 def test_getting_authenticated_current_user_success(app_with_test_user, test_user, client: Client):
     client = app_with_test_user.test_client()
-    
-    # Save to database within app context
-    with app_with_test_user.app_context():        
-        # Create token with the user's ID as identity
+    with app_with_test_user.app_context():
         token = create_access_token(identity="foo")
-    
-    # Test the endpoint
+
     response = client.get("/users/me", headers={"Authorization": f"Bearer {token}"})
-    
-    # Verify successful response
     assert response.status_code == 200
-    
-    # Verify response data (adjust according to your API's response format)
     response_data = response.get_json()
     assert response_data["name"] == test_user.name
     assert response_data["email"] == test_user.email
 
 def test_update_current_user(app_with_test_user, client: Client):
     client = app_with_test_user.test_client()
-    
-    # Save to database within app context
-    with app_with_test_user.app_context():        
-        # Create token with the user's ID as identity
+    with app_with_test_user.app_context():
         token = create_access_token(identity="foo")
-    
-    # Test the endpoint
     response = client.put("/users/me", headers={"Authorization": f"Bearer {token}"}, json={"email": "foo@bar.qux", "name": "foo"})
-    
-    # Verify successful response
     assert response.status_code == 200, response.get_data()
-    
-    # Verify response data (adjust according to your API's response format)
     response_data = response.get_json()
     assert response_data["name"] == "foo"
     assert response_data["email"] == "foo@bar.qux"
